LegalDOCAI/requirements.txt
# FastAPI backend
fastapi==0.95.2
uvicorn[standard]==0.23.2

# Database
pymongo==4.4.0

# File handling & uploads
python-multipart==0.0.6
python-dotenv==1.0.0

# Image & OCR
Pillow==10.4.0
pytesseract==0.3.13

# PDF handling
PyMuPDF==1.26.4

# Machine Learning & NLP
numpy==1.26.4
scikit-learn==1.3.2
spacy==3.8.1
sentence-transformers==2.2.2
transformers==4.34.0
torch==2.2.0
